+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * RDSActionLabel can be used from Objective-C
 * Support multi-word matchers
+* Respect label's text alignment and line break mode
 
 ## Version 0.1.0